Move base class for Image and Labels to own file
