+++ napari/layers/intensity_mixin.py
@@ -11,7 +11,7 @@ from napari.utils.validators import _validate_increasing, validate_n_seq
 validate_2_tuple = validate_n_seq(2)
 
 if TYPE_CHECKING:
-    from napari.layers.image.image import _ImageBase
+    from napari.layers._scalar_field.scalar_field import _ScalarFieldBase
 
 
 class IntensityVisualizationMixin:
@@ -49,7 +49,7 @@ class IntensityVisualizationMixin:
         self._auto_contrast_source = 'slice'
         self._keep_auto_contrast = False
 
-    def reset_contrast_limits(self: '_ImageBase', mode=None):
+    def reset_contrast_limits(self: '_ScalarFieldBase', mode=None):
         """Scale contrast limits to data range"""
         mode = mode or self._auto_contrast_source
         self.contrast_limits = self._calc_data_range(mode)
